+++ index.ts
@@ -8,6 +8,7 @@ import { origin, port, secret } from './env';
 import schema from './schema';
 import { MyContext } from './context';
 import sql from 'sql-template-strings';
+import { UnsplashApi } from './schema/unsplash.api';
 const { PostgresPubSub } = require('graphql-postgres-subscriptions');
 
 const pubsub = new PostgresPubSub({
@@ -67,6 +68,9 @@ const server = new ApolloServer({
 
     return res;
   },
+  dataSources: () => ({
+    unsplashApi: new UnsplashApi(),
+  }),
 });
 
 server.applyMiddleware({
